# hangsha/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Event Calendar API
  version: 1.0.0
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    GetMeResponse:
      type: object
      required: [id]
      properties:
        id: { type: integer, format: int64 }
        username: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        profileImageUrl: { type: string, nullable: true }
...
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      properties:
        username:
          type: string
          nullable: true
          description: |
            닉네임 변경.
            - 필드 미포함: 변경 없음
            - null: 값 비우기(삭제)
            - ""(빈 문자열): 서버는 빈 문자열도 값으로 저장합니다(현재 코드 기준).
        profileImageUrl:
          type: string
          nullable: true
          format: uri
          pattern: '^https?://.*'   # http/https만 허용
          description: |
            프로필 이미지 URL 변경.
            - 필드 미포함: 변경 없음
            - null: 값 비우기(삭제)
            - ""(빈 문자열): 서버에서 null로 처리(삭제)
      description: |
        내 프로필 PATCH.
        요청 바디는 username/profileImageUrl 중 하나 이상을 포함해야 함.

    # ---------------- Categories ----------------
...
    AddExcludedKeywordRequest:
      type: object
      required: [keyword]
      properties:
        keyword: { type: string, minLength: 1 }
    ListExcludedKeywordResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [id, keyword, createdAt]
            properties:
              id: { type: integer, format: int64 }
              keyword: { type: string }
              createdAt: { type: string, format: date-time }
    ReplaceAllInterestCategoriesRequest:
      type: object
      required: [ items ]
      properties:
        items:
          type: array
          minItems: 0
          description: "유저의 관심 카테고리 전체 목록을 이 배열로 교체합니다."
          items:
            type: object
            required: [ categoryType, categoryId, priority ]
            properties:
              categoryType: { type: string, enum: [EVENT_STATUS, EVENT_TYPE, ORGANIZATION] }
              categoryId: { type: integer, format: int64 }
              priority:
                type: integer
                minimum: 1
                description: "1이 가장 높음. 권장: 1..N 연속, 중복 없음"
...
    ListInterestCategoryResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [categoryType, categoryId, name, sortOrder, priority]
            properties:
              categoryType: { type: string, enum: [EVENT_STATUS, EVENT_TYPE, ORGANIZATION] }
              categoryId: { type: integer, format: int64 }
              name: { type: string }
              sortOrder: { type: integer }
              priority:
                type: integer
                minimum: 1
                description: "유저 관심 카테고리 우선순위 (1이 가장 높음)"

    # ---------------- Events ----------------
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: 내 정보 조회
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GetMeResponse' }

    patch:
      tags: [ Users ]
      summary: 내 프로필 수정 (username / profileImageUrl)
      description: |
        - 필드 미포함: 변경 없음
        - null: 해당 값 비우기(삭제)
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProfileRequest' }
            examples:
              setUsername:
                value: { username: "dyang" }
              clearUsername:
                value: { username: null }
              clearProfileImage:
                value: { profileImageUrl: null }
      responses:
        "204": { description: No Content }
        "400":
          description: Bad Request
            - username/profileImageUrl 둘 다 없음
            - profileImageUrl이 유효한 http/https URL이 아님
            - username이 금칙어/예약어에 해당함 (USERNAME_NOT_ALLOWED)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ---------------- Categories ----------------
...
  /users/me/excluded-keywords:
    get:
      tags: [Preference]
      summary: 제외 키워드 목록
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListExcludedKeywordResponse' }
    post:
      tags: [Preference]
      summary: 제외 키워드 추가
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddExcludedKeywordRequest' }
      responses:
        "201": { description: Created }
  /users/me/interest-categories:
    get:
      tags: [Preference]
      summary: 관심 카테고리 목록 (우선순위 포함)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListInterestCategoryResponse' }

    put:
      tags: [ Preference ]
      summary: 관심 카테고리 전체 교체(배치 저장)
      description: |
        유저의 관심 카테고리 구성을 요청 바디의 items로 "전체 교체"합니다.
        - items가 빈 배열이면 관심 카테고리를 모두 삭제합니다.
        - 권장 규칙: priority는 1..N 연속 + 중복 없음 (서버에서 검증)
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceAllInterestCategoriesRequest' }
      responses:
        "204": { description: No Content }
        "400":
          description: Bad Request (중복 categoryId/priority, priority 규칙 위반 등)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/me/interest-categories/{categoryId}:
    delete:
      tags: [Preference]
      summary: 관심 카테고리 삭제
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: categoryId
          required: true
          schema: { type: integer, format: int64 }
        - in: query
          name: categoryType
          required: true
          schema: { type: string, enum: [EVENT_STATUS, EVENT_TYPE, ORGANIZATION] }
      responses:
        "204": { description: No Content }

  # ---------------- Events ----------------
...